Backend/service/ai-service/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
pillow==10.2.0
numpy==1.26.3
torch==2.1.2
torchvision==0.16.2
redis==5.0.1
supabase==2.3.0
python-dotenv==1.0.1
pydantic==2.5.3
starlette==0.35.1     
python-jose[cryptography]==3.3.0
opencv-python-headless==4.9.0.80
requests==2.31.0
google-generativeai==0.8.6
tqdm==4.66.1
typing-extensions==4.9.0
urllib3==2.1.0
websockets==12.0
uvloop==0.19.0 ; sys_platform != "win32"
watchfiles==0.21.0
httptools==0.6.1
loguru==0.7.2
matplotlib==3.8.2
pandas==2.2.0
scikit-learn==1.4.0
scipy==1.12.0
seaborn==0.13.1
statsmodels==0.14.1
xlrd==2.0.1
XlsxWriter==3.1.9
PyYAML==6.0.1
Jinja2==3.1.3
MarkupSafe==2.1.4
click==8.1.7
colorama==0.4.6
tqdm==4.66.1
regex==2023.12.25
requests-toolbelt==1.0.0
sniffio==1.3.0
soupsieve==2.5
tomli==2.0.1
urllib3==2.1.0
webencodings==0.5.1
websocket-client==1.7.0
zipp==3.17.0
poetry==1.7.1
poetry-core==1.8.1
poetry-plugin-export==1.6.0
virtualenv==20.25.0
pip==23.3.2 
setuptools==69.0.3
wheel==0.42.0

